Move nemo and openai fixtures up a level to be shared with benchmarks
